RF: Move ria_utils to customremotes
